Re-factor auths testing
